+++ src/util.rs
@@ -1,0 +1,33 @@
+pub fn as_readable_time(dur: &std::time::Duration) -> String {
+    const TABLE: [(&str, u64); 4] = [
+        ("days", 86400),
+        ("hours", 3600),
+        ("minutes", 60),
+        ("seconds", 1),
+    ];
+
+    fn pluralize(s: &&str, n: u64) -> String {
+        format!("{} {}", n, if n > 1 { s } else { &s[..s.len() - 1] })
+    }
+
+    let mut time = vec![];
+    let mut secs = dur.as_secs();
+    for (name, d) in &TABLE {
+        let div = secs / d;
+        if div > 0 {
+            time.push(pluralize(name, div));
+            secs -= d * div;
+        }
+    }
+
+    let len = time.len();
+    if len > 1 {
+        if len > 2 {
+            for e in &mut time.iter_mut().take(len - 2) {
+                e.push_str(",")
+            }
+        }
+        time.insert(len - 1, "and".into());
+    }
+    time.join(" ")
+}
